requirements.txt
appdirs==1.4.3
Django==3.1.7
djangorestframework==3.12.4
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0